back/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0.post1
python-dateutil==2.8.2